lambdas/data_loader/requirements.txt
boto3==1.28.38
psycopg2-binary==2.9.6 